refactor: [M3-6375] - MUI v5 - Components > PreferenceToggle
