--- packages/manager/src/components/PreferenceToggle/index.ts
@@ -1,6 +1,0 @@
-import { ToggleProps as _ToggleProps } from './PreferenceToggle';
-
-/* tslint:disable */
-export interface ToggleProps<T> extends _ToggleProps<T> {}
-
-export { default } from './PreferenceToggle';
